pyproject.toml
[project]
name = "filter-openai-plugin"
version = "0.1.0"
description = "OpenAI filter plugin for photobooth-app"
authors = ["Your Name <your@email.com>"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Pillow",
    "niquests",
    "pydantic",
    "pydantic-settings",
]

[tool.pytest.ini_options]
pythonpath = ["src", "photobooth-data/plugins"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest",
    "photobooth-app",
]

[tool.ruff]
line-length = 150
target-version = "py310"